+++ virtcontainers/utils/utils_linux_test.go
@@ -15,7 +15,7 @@ import (
 func TestFindContextID(t *testing.T) {
 	assert := assert.New(t)
 
-	ioctlFunc = func(fd uintptr, request int, arg1 uint32) error {
+	ioctlFunc = func(fd uintptr, request int, arg1 uint64) error {
 		return errors.New("ioctl")
 	}
 
@@ -26,7 +26,7 @@ func TestFindContextID(t *testing.T) {
 		maxUInt = orgMaxUInt
 	}()
 	VHostVSockDevicePath = "/dev/null"
-	maxUInt = uint32(1000000)
+	maxUInt = uint64(1000000)
 
 	f, cid, err := FindContextID()
 	assert.Nil(f)
